+++ src/components/appbar/__test__/index.config.ts
@@ -1,0 +1,5 @@
+export default {
+  navigationBarTitleText: "AppBar",
+  navigationStyle: "custom",
+  navigationBarTextStyle: "black",
+};