+++ src/pages/index/index.config.ts
@@ -1,5 +1,3 @@
 export default {
   navigationBarTitleText: "首页",
-  // navigationStyle: "custom",
-  // navigationBarTextStyle: "white",
 };