+++ src/pages/transition/index.config.ts
@@ -1,0 +1,3 @@
+export default {
+  navigationBarTitleText: "过渡动画",
+};